ENH: Add Event Receiver Class
Add class to receive event.
Provides the IsKeyDown( key_param ) method to check if
the key parameter key_param is pressed.
The behavior to adopt when a key is pressed is still
defined in main.cpp. In this case, we add the possibility
to close the window if the escape key is pressed.