NEW: Add Path Finder
Add PathFinder class to return full path to files stored
in the source directory. This class uses the Project_Config.h
to access global variable such as the path to the media directory.
Call getFullMediaPath( filename ) on a PathFinder object to get
the full path to the file called filename in the media directory.