NEW: Add Camera Manager
Add Camera manager to add a camera in the scene. This commit
also add collision handling between the terrain and the camera.
The following approach has been chosen to handle collision :
For each object as a scene node, provide the different scene
node it has to collide with as an input parameter of the
addSceneNodeCollision() method.