BUG: Fix character mesh offset
Add offset to the terrainManager collision function to
improve mesh-terrain collision. The offset is hardcoded
and depends on the origin of the mesh. it has to be
adjusted so the lowest point of the mesh touch the terrain.

This commit also improves the 3rd person camera rotation
by using Irrlicht rotate function.